requirements.txt
notion-client>=2.0.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
schedule>=1.2.0
streamlit>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0